backend/requirements.txt
Flask==3.0.0
Flask-SocketIO==5.3.6
Flask-SQLAlchemy==3.1.1
Flask-Cors==4.0.0
Flask-Limiter==3.5.0
Flask-Caching==2.1.0
firebase-admin==6.2.0
google-cloud-firestore
google-cloud-logging
google-cloud-storage
google-cloud-secret-manager
google-auth
google-generativeai
flask-talisman
bleach
pytest==7.4.2
python-dotenv
gunicorn
eventlet
